Move events recorder to runtime
