--- recorder/go.mod
@@ -1,10 +1,0 @@
-module github.com/fluxcd/pkg/recorder
-
-go 1.14
-
-require (
-	github.com/hashicorp/go-retryablehttp v0.6.6
-	github.com/stretchr/testify v1.4.0
-	k8s.io/api v0.18.8
-	k8s.io/apimachinery v0.18.8
-)
